ref: Address review feedback
