+++ relay-cabi/src/processing.rs
@@ -8,7 +8,7 @@ use std::slice;
 
 use json_forensics;
 use relay_common::{glob_match_bytes, GlobOptions};
-use relay_general::pii::{selectors_from_event, DataScrubbingConfig, PiiConfig, PiiProcessor};
+use relay_general::pii::{selectors_from_value, DataScrubbingConfig, PiiConfig, PiiProcessor};
 use relay_general::processor::{process_value, split_chunks, ProcessingState};
 use relay_general::protocol::{Event, VALID_PLATFORMS};
 use relay_general::store::{GeoIpLookup, StoreConfig, StoreProcessor};
@@ -183,8 +183,8 @@ ffi_fn! {
     /// Walk through the event and collect selectors that can be applied to it in a PII config. This
     /// function is used in the UI to provide auto-completion of selectors.
     unsafe fn relay_pii_selectors_from_event(event: *const RelayStr) -> Result<RelayStr> {
-        let event = Annotated::<Event>::from_json((*event).as_str())?;
-        let rv = selectors_from_event(event);
+        let mut event = Annotated::<Event>::from_json((*event).as_str())?;
+        let rv = selectors_from_value(&mut event);
         Ok(RelayStr::from_string(serde_json::to_string(&rv)?))
     }
 }
